join: exact-size window-join right-column index (pre-existing >256-col overflow)
exec_window_join's right_out_idx[] was a fixed int64_t[256] filled by a
loop over the right table's total column count (right_ncols), skipping
only the time key and eq keys. Nothing capped right_ncols, so an
asof/window join whose right table has >256 columns overflowed it.
ASan-confirmed by review: a 260-column right table (Time + d001..d259)
joined via `asof-join [Time]` produces stack-buffer-overflow WRITE in
exec_window_join, join.c:3029 (right_out_count=259 entries into a
256-slot array). Fixed by carving right_out_idx from scratch_alloc,
sized exactly to right_ncols, freed at all three exits at/below the
declaration (its own OOM guard, the existing lidx/ridx OOM guard, and
the function's final success return).

Also: tblop.c:998's `ray_distinct(g, keys, (uint8_t)ncols)` narrowed
into a uint8_t despite ray_distinct's uint32_t signature -- the exact
truncating-carrier pattern this migration retires (safe today via the
ncols>255 check above, but the cast itself should not exist). Widened
to (uint32_t)ncols.

Regression test added to test/rfl/query/width_matrix.rfl: builds a
260-column right table entirely in-harness (bracket name literal +
take/concat-replicated columns, no /tmp fixtures) and asserts the
value of the last column (past the old 256 boundary) survives the
asof join. Verified the committed test crashes pre-fix (ASan, exactly
at join.c:3029) by stashing only the join.c change and rebuilding.

+++ src/ops/tblop.c
@@ -995,7 +995,7 @@ ray_t* ray_table_distinct_fn(ray_t* tbl) {
         if (!keys[c]) { ray_graph_free(g); return ray_error("oom", NULL); }
     }
 
-    ray_op_t* root = ray_distinct(g, keys, (uint8_t)ncols);
+    ray_op_t* root = ray_distinct(g, keys, (uint32_t)ncols);
     if (!root) { ray_graph_free(g); return ray_error("oom", NULL); }
 
     ray_t* result = ray_execute(g, root);
